refactor:(UI module) — fix narrowing conversions via explicit casts
— UI/wxWidgets module only (no module mixing).
Behavior-preserving: every change is an explicit static_cast<> that
documents an already-implicit conversion. No runtime behavior change.
Verified with a full MSVC/Ninja build (links OK).

44 narrowing sites fixed across 11 UI files:
- browse_tile_window, editor_tabs, main_toolbar, properties_window:
  size_t/size_type -> int for displayed counts and UI control indices
- result_window, palette_brushlist, palette_creature, palette_house,
  common_windows: unsigned/size_t -> int for list SetSelection/IsSelected,
  keyboard list navigation, Append client-data, wxSpinCtrl init value
- old_properties_window: Append client-data, GetCount()-1 index
- main_menubar: SetLoadDone(int32_t) progress (x12), recent-files event ID,
  displayed town_count/house_count

Deliberately NOT changed (out of UI scope / "when in doubt, don't"):
- gui.cpp brush-thickness draw math; map_window DPI coordinate scaling
- container_properties iterator arithmetic; replace_items map operation
- ContainerItemButton slot index; preferences float-widening arithmetic

No rendering/OpenGL/map-coordinate/serialization/engine logic touched.
No base types, classes, algorithms or structure changed.

+++ source/browse_tile_window.cpp
@@ -154,13 +154,13 @@ BrowseTileWindow::BrowseTileWindow(wxWindow* parent, Tile* tile, wxPoint positio
 	infoSizer->Add(buttons);
 	infoSizer->AddSpacer(5);
 	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "Position:  " + pos), wxSizerFlags(0).Left());
-	infoSizer->Add(item_count_txt = newd wxStaticText(this, wxID_ANY, "Item count:  " + i2ws(item_list->GetItemCount())), wxSizerFlags(0).Left());
+	infoSizer->Add(item_count_txt = newd wxStaticText(this, wxID_ANY, "Item count:  " + i2ws(static_cast<int>(item_list->GetItemCount()))), wxSizerFlags(0).Left());
 	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "Protection zone:  " + b2yn(tile->isPZ())), wxSizerFlags(0).Left());
 	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "No PvP:  " + b2yn(tile->getMapFlags() & TILESTATE_NOPVP)), wxSizerFlags(0).Left());
 	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "No logout:  " + b2yn(tile->getMapFlags() & TILESTATE_NOLOGOUT)), wxSizerFlags(0).Left());
 	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "PvP zone:  " + b2yn(tile->getMapFlags() & TILESTATE_PVPZONE)), wxSizerFlags(0).Left());
 	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "House:  " + b2yn(tile->isHouseTile())), wxSizerFlags(0).Left());
-	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "Zones:  " + i2ws(tile->zones.size())), wxSizerFlags(0).Left());
+	infoSizer->Add(newd wxStaticText(this, wxID_ANY, "Zones:  " + i2ws(static_cast<int>(tile->zones.size()))), wxSizerFlags(0).Left());
 
 	sizer->Add(infoSizer, wxSizerFlags(0).Left().DoubleBorder());
 
@@ -189,7 +189,7 @@ void BrowseTileWindow::OnItemSelected(wxCommandEvent& WXUNUSED(event)) {
 
 void BrowseTileWindow::OnClickDelete(wxCommandEvent& WXUNUSED(event)) {
 	item_list->RemoveSelected();
-	item_count_txt->SetLabelText("Item count:  " + i2ws(item_list->GetItemCount()));
+	item_count_txt->SetLabelText("Item count:  " + i2ws(static_cast<int>(item_list->GetItemCount())));
 }
 
 void BrowseTileWindow::OnClickSelectRaw(wxCommandEvent& WXUNUSED(event)) {
