+++ source/palette_brushlist.cpp
@@ -586,7 +586,7 @@ Brush* BrushListBox::GetSelectedBrush() const {
 bool BrushListBox::SelectBrush(const Brush* whatbrush) {
 	for (size_t n = 0; n < tileset->size(); ++n) {
 		if (tileset->brushlist[n] == whatbrush) {
-			SetSelection(n);
+			SetSelection(static_cast<int>(n));
 			return true;
 		}
 	}
